pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chile-balance"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv",
    "requests",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "responses",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/chile_balance"]
